commented error report of compass to reduce memory usage

+++ arduino_motor/mains/HMC5883LCompass.h
@@ -1,6 +1,5 @@
 #ifndef HMC5883LCompass_H
 #define HMC5883LCompass_H
-
 #include <Wire.h>
 #include <HMC5883L.h>
 class HMC5883LCompass {
@@ -10,7 +9,7 @@ private:
   void setupHMC5883L();
 public:
   HMC5883LCompass();
-  double getHeading();
+  float getHeading();
 };
 
 #endif
